requirements.txt
anthropic>=0.54.0
mcp>=1.0.0
pydantic>=2.8.0
python-dotenv>=1.0.1
pytest>=8.3.0